pyscript/requirements.txt
# required packages
digi-xbee